Performance 1.5x, memory used 1/2
