+++ lib/mutils.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require 'dry/inflector'
-require 'oj'
 require_relative 'mutils/version'
 require_relative 'mutils/lib/helper'
+require_relative 'mutils/lib/result_hash'
 require_relative 'mutils/serialization/serialization_results'
 require_relative 'mutils/serialization/serialization_includes'
 require_relative 'mutils/serialization/serialization_methods'
