+++ spec/spec_helper.rb
@@ -3,6 +3,7 @@
 require 'bundler/setup'
 require 'mutils'
 require 'coveralls'
+require 'benchmark'
 require 'rspec-benchmark'
 
 Coveralls.wear!
@@ -13,6 +14,9 @@ RSpec.configure do |config|
   config.disable_monkey_patching!
   config.expect_with :rspec do |c|
     c.syntax = :expect
+  end
+  if ENV['PERFORMANCE_ONLY'] == 'true' || ENV['PERFORMANCE_ONLY'] == true
+    config.filter_run_excluding general: true
   end
   if ENV['TRAVIS'] == 'true' || ENV['TRAVIS'] == true
     config.filter_run_excluding performance: true
